requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.19
pydantic==2.10.0
pydantic-settings==2.6.0

# PDF Processing
pdfplumber==0.11.4
PyMuPDF==1.24.14
pillow==11.0.0

# DOCX Generation
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
httpx==0.28.1

# Development
black==24.10.0
ruff==0.8.4
mypy==1.13.0
